Append infra models and apis

+++ web/src/app/api/infra/[id]/route.ts
@@ -1,6 +1,8 @@
 import { NextRequest, NextResponse } from 'next/server';
 import dbConnect from '@/lib/mongoose';
 import Infra from '@/lib/models/Infra';
+import { getSession } from '@/lib/session';
+import { Role } from '@/lib/models/User';
 
 
 /**
@@ -99,23 +101,26 @@ import Infra from '@/lib/models/Infra';
 export async function PUT(req: NextRequest, { params }: { params: { id: string } }) {
   await dbConnect();
   try {
+    const session = await getSession();
+    if (!session || !session.isAuth) {
+      return NextResponse.json({ success: false, error: "Authentication required" }, { status: 401 })
+    }
+    if (session.role != Role.ADMIN) {
+      return NextResponse.json({ success: false, error: "Permission required" }, { status: 403 })
+    }
     const { id } = params;
     const { title, description, url } = await req.json();
-
     if (!title || !description || !url) {
-      return NextResponse.json({ message: "Title, Description, and URL are required." }, { status: 400 });
+      return NextResponse.json({ success: false, error: "Title, Description, and URL are required." }, { status: 400 });
     }
-
     const updatedInfra = await Infra.findByIdAndUpdate(id, { title, description, url }, { new: true });
-
     if (!updatedInfra) {
-      return NextResponse.json({ message: "Infra not found" }, { status: 404 });
+      return NextResponse.json({ success: false, error: "Infra not found" }, { status: 404 });
     }
-
-    return NextResponse.json({ message: "Infra updated successfully", infra: updatedInfra });
+    return NextResponse.json({ success: true });
   } catch (error) {
     console.error('Error updating infra:', error);
-    return NextResponse.json({ message: "Server error occurred" }, { status: 500 });
+    return NextResponse.json({ success: false, error: "Server error occurred" }, { status: 500 });
   }
 }
 
@@ -170,16 +175,22 @@ export async function PUT(req: NextRequest, { params }: { params: { id: string }
 export async function DELETE(req: NextRequest, { params }: { params: { id: string } }) {
   await dbConnect();
   try {
+    const session = await getSession();
+    if (!session || !session.isAuth) {
+      return NextResponse.json({ success: false, error: "Authentication required" }, { status: 401 })
+    }
+    if (session.role != Role.ADMIN) {
+      return NextResponse.json({ success: false, error: "Permission required" }, { status: 403 })
+    }
     const { id } = params;
     const deletedInfra = await Infra.findByIdAndDelete(id);
 
     if (!deletedInfra) {
-      return NextResponse.json({ message: "Infra not found" }, { status: 404 });
+      return NextResponse.json({ success: false, error: "Infra not found" }, { status: 404 });
     }
-
-    return NextResponse.json({ message: "Infra deleted successfully" });
+    return NextResponse.json({ success: true });
   } catch (error) {
     console.error('Error deleting infra:', error);
-    return NextResponse.json({ message: "Server error occurred" }, { status: 500 });
+    return NextResponse.json({ success: false, error: "Internal server error" }, { status: 500 });
   }
 }
